refactor: rename functions inside updateStyle, move function to separate object
